pyproject.toml
[project]
name = "djangoproject2"
version = "0.1.0"
description = ""
authors = [
    {name = "pyoneerC",email = "maxcomperatutti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
